Move unsupported types test to its own spec
